requirements.txt
jiwer
pymupdf
requests
python-dotenv
google-genai